requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==1.26.4
openpyxl==3.1.2
pandas==2.1.1
psycopg2-binary==2.9.10
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
Werkzeug==3.1.3
Flask-APScheduler==1.12.4
Flask-Mail==0.10.0
